02 test saga bugfix

+++ services/authentication/src/main/java/com/graduationProject/authentication/saga/sagaParticipators/DeleteResident.java
@@ -8,6 +8,7 @@ import com.graduationProject.authentication.kafka.KafkaApi;
 import com.graduationProject.authentication.saga.SagaParticipator;
 import com.graduationProject.authentication.service.ResidentService;
 import com.graduationProject.authentication.type.SagaStatus;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -29,14 +30,17 @@ public class DeleteResident implements SagaParticipator<SagaResidentDto> {
     public void transact(SagaResidentDto sagaResidentDto) {
         SagaResponseDto sagaResponseDto;
         try {
-            if (sagaResidentDto.getResidentDto().getUsername().equals("neverDelete")) {
-                throw new IllegalArgumentException("Admin can never be deleted");
+            if (sagaResidentDto.getResidentDto().getUsername().equals("neverDelete") || sagaResidentDto.getResidentDto().getUsername().equals("bFailRevert")) {
+                throw new IllegalStateException(String.format("Could not delete Resident with \n ID: %s \n SagaID: %s",
+                        sagaResidentDto.getResidentDto().getId(),
+                        sagaResidentDto.getSagaId()));
             }
             residentService.deleteResident(sagaResidentDto.getResidentDto().getId());
             sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.SUCCESS);
 
         } catch (Exception e) {
             sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.FAILED);
+            sagaResponseDto.setErrorMessage(ExceptionUtils.getStackTrace(e));
             e.printStackTrace();
         }
 
@@ -47,20 +51,21 @@ public class DeleteResident implements SagaParticipator<SagaResidentDto> {
     public void revert(SagaResidentDto sagaResidentDto) {
         SagaResponseDto sagaResponseDto;
         try {
-            if (sagaResidentDto.getResidentDto().getPassword().equals("neverDelete")) {
-                throw new Exception();
+            if (sagaResidentDto.getResidentDto().getPassword().equals("neverDelete") ) {
+                throw new IllegalStateException(String.format("Could not revert deletion of Resident with \n ID: %s \n SagaID: %s",
+                        sagaResidentDto.getResidentDto().getId(),
+                        sagaResidentDto.getSagaId()));
             }
+            residentService.deleteIfExists(sagaResidentDto.getResidentDto().getId());
             residentService.addResident(sagaResidentDto.getResidentDto());
             sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.SUCCESS);
         } catch (Exception e) {
             sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.FAILED);
+            sagaResponseDto.setErrorMessage(ExceptionUtils.getStackTrace(e));
             e.printStackTrace();
         }
-
         produceKafkaMessage(DeleteResidentSagaRevert, sagaResponseDto);
-
     }
-
 
     private void produceKafkaMessage(String residentTopic, SagaResponseDto sagaResponseDto) {
         try {
@@ -69,5 +74,4 @@ public class DeleteResident implements SagaParticipator<SagaResidentDto> {
             e.printStackTrace();
         }
     }
-
 }
